refactor: Added Resize transforms for PyTorch
