+++ doctr/models/preprocessor/pytorch.py
@@ -10,43 +10,9 @@ from typing import List, Tuple, Union, Any
 
 from torchvision.transforms import transforms as T
 from torchvision.transforms import functional as F
-from torch.nn.functional import pad
+from doctr.transforms import Resize
 
 __all__ = ['PreProcessor']
-
-
-class Resize(T.Resize):
-    def __init__(
-        self,
-        size: Tuple[int, int],
-        interpolation=F.InterpolationMode.BILINEAR,
-        preserve_aspect_ratio: bool = False,
-        symmetric_pad: bool = False,
-    ) -> None:
-        super().__init__(size, interpolation)
-        self.preserve_aspect_ratio = preserve_aspect_ratio
-        self.symmetric_pad = symmetric_pad
-
-    def forward(self, img: torch.Tensor) -> torch.Tensor:
-        target_ratio = self.size[0] / self.size[1]
-        actual_ratio = img.shape[0] / img.shape[1]
-        if not self.preserve_aspect_ratio or (target_ratio == actual_ratio):
-            return super().forward(img)
-        else:
-            if actual_ratio > target_ratio:
-                tmp_size = (self.size[0], int(self.size[0] / actual_ratio))
-            else:
-                tmp_size = (int(self.size[1] * actual_ratio), self.size[1])
-            # Scale image
-            img = F.resize(img, tmp_size, self.interpolation)
-            # Pad
-            _pad = (self.size[0] - img.shape[0], self.size[1] - img.shape[1])
-            if self.symmetric_pad:
-                half_pad = (math.ceil(_pad[0] / 2), math.ceil(_pad[1] / 2))
-                _pad = (half_pad[0], _pad[0] - half_pad[0], half_pad[1], _pad[1] - half_pad[1])
-            else:
-                _pad = (0, _pad[0], 0, _pad[1])
-            return pad(img, _pad)
 
 
 class PreProcessor(nn.Module):
@@ -114,12 +80,12 @@ class PreProcessor(nn.Module):
         # Check input type
         if isinstance(x, torch.Tensor):
             # Tf tensor from data loader: check if tensor size is output_size
-            if x.shape[1] != self.resize.output_size[0] or x.shape[2] != self.resize.output_size[1]:
-                x = F.resize(x, self.resize.output_size, method=self.resize.method)
+            if x.shape[1] != self.resize.size[0] or x.shape[2] != self.resize.size[1]:
+                x = F.resize(x, self.resize.size, interpolation=self.resize.interpolation)
             processed_batches = [x]
         elif isinstance(x, list):
             # convert images to tf
-            tensors = [sample.to(dtype=torch.float32) for sample in x]
+            tensors = [torch.from_numpy(sample) for sample in x]
             # Resize (and eventually pad) the inputs
             images = [self.resize(sample) for sample in tensors]
             # Batch them
