+++ doctr/transforms/functional/__init__.py
@@ -1,0 +1,4 @@
+from doctr.file_utils import is_tf_available
+
+if is_tf_available():
+    from .tensorflow import *
