+++ doctr/transforms/functional/tensorflow.py
