+++ doctr/transforms/modules/pytorch.py
@@ -1,0 +1,55 @@
+# Copyright (C) 2021, Mindee.
+
+# This program is licensed under the Apache License version 2.
+# See LICENSE or go to <https://www.apache.org/licenses/LICENSE-2.0.txt> for full license details.
+
+import math
+import torch
+from torchvision.transforms import transforms as T
+from torchvision.transforms import functional as F
+from torch.nn.functional import pad
+from typing import Tuple
+
+
+__all__ = ['Resize']
+
+
+class Resize(T.Resize):
+    def __init__(
+        self,
+        size: Tuple[int, int],
+        interpolation=F.InterpolationMode.BILINEAR,
+        preserve_aspect_ratio: bool = False,
+        symmetric_pad: bool = False,
+    ) -> None:
+        super().__init__(size[::-1], interpolation)
+        self.preserve_aspect_ratio = preserve_aspect_ratio
+        self.symmetric_pad = symmetric_pad
+
+    def forward(self, img: torch.Tensor) -> torch.Tensor:
+        target_ratio = self.size[0] / self.size[1]
+        actual_ratio = img.shape[-2] / img.shape[-1]
+        if not self.preserve_aspect_ratio or (target_ratio == actual_ratio):
+            return super().forward(img)
+        else:
+            if actual_ratio > target_ratio:
+                tmp_size = (self.size[1], int(self.size[1] / actual_ratio))
+            else:
+                tmp_size = (int(self.size[0] * actual_ratio), self.size[0])
+            # Scale image
+            img = F.resize(img, tmp_size, self.interpolation)
+            # Pad (inverted in pytorch)
+            _pad = (self.size[1] - img.shape[-1], self.size[0] - img.shape[-2])
+            if self.symmetric_pad:
+                half_pad = (math.ceil(_pad[0] / 2), math.ceil(_pad[1] / 2))
+                _pad = (half_pad[0], _pad[0] - half_pad[0], half_pad[1], _pad[1] - half_pad[1])
+            else:
+                _pad = (0, _pad[0], 0, _pad[1])
+            return pad(img, _pad)
+
+    def __repr__(self) -> str:
+        interpolate_str = self.interpolation.value
+        _repr = f"output_size={self.size[::-1]}, interpolation='{interpolate_str}'"
+        if self.preserve_aspect_ratio:
+            _repr += f", preserve_aspect_ratio={self.preserve_aspect_ratio}, symmetric_pad={self.symmetric_pad}"
+        return f"{self.__class__.__name__}({_repr})"
